+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 ### latest
 
 **Added**:
- - Consumer middleware
+ - Message decode logic is moved to a middleware
 
 **Updated**
- - Context obj new field 'decoded_value'
+ - Context obj new field 'decoded_value_cb'
+ - **breaking**: request.json is now awaitable
 
 ### 0.2.0 (2023-10-14)